fix(terminal): auto-focus newly created terminals
Cmd+T, the "+" dropdown, and the auto-created first terminal on a
freshly activated worktree now leave keyboard focus inside the new
xterm so typing works without a click.

Previously only the split-group "+" dropdown's onCloseAutoFocus path
attempted this, and its single double-rAF could still lose to WebGL
setup/Radix restore on some machines. Extracted a shared
focusNewTerminalTab helper that polls across rAFs until the new
xterm-helper-textarea is attached, and wired it into the Cmd+T /
legacy TabBar handler and the workspace-activation auto-create effect.
